requirements.txt
boto3
openpyxl
pandas
redshift_connector
python-dotenv
pylint
pytest
pyarrow
altair
argparse
streamlit
watchdog
datetime